requirements.txt
requests>=2.31.0
psycopg2-binary>=2.9.0
pyyaml>=6.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pdfplumber>=0.10.0
feedparser>=6.0.10
